refactor(elevenlabs): lift the panel presets out of TtsExampleCard
TtsExampleCard is shared by every TTS provider, and the ElevenLabs work had
grown it from 612 to 1313 lines. The v3 tag palette, Enhance vibes and
director prompt, language roster, starter scripts and the settings row are
pure data plus one presentational component, so they move to
elevenlabsPanel.js and the card imports them.

No behaviour change: the card keeps the state and handlers, only module-level
constants moved. Verified against a live ElevenLabs key before and after —
same rendered panel, same generated audio, same history entry.
